moving UI components and containers into UI folder
