--- src/js/modules/auth/components/index.js
@@ -1,4 +1,0 @@
-export LoginForm from './LoginForm';
-export EmailInput from './EmailInput';
-export PasswordInput from './PasswordInput';
-export LoginView from './LoginView';
