--- src/js/modules/auth/containers/index.js
@@ -1,4 +1,0 @@
-import connectToLogin from './LoginContainer';
-
-export const connectToLoginContainer = connectToLogin;
-export default connectToLoginContainer;
