+++ src/js/modules/auth/ui/components/PasswordInput.js
