+++ src/js/modules/auth/ui/index.js
@@ -1,0 +1,10 @@
+import EmailInput from './components/EmailInput';
+import LoginForm from './components/LoginForm';
+import LoginView from './components/LoginView';
+import LogoutButton from './components/LogoutButton';
+import PasswordInput from './components/PasswordInput';
+
+import LoginContainer from './containers/LoginContainer';
+
+export const components = { EmailInput, LoginForm, LoginView, LogoutButton, PasswordInput };
+export const containers = { LoginContainer };
